moved Saml2Login API view to separate file
