+++ poem/Poem/api/internal_views/login.py
@@ -1,0 +1,34 @@
+from django.core.cache import cache
+
+from Poem.poem.saml2.config import get_schemaname
+
+from rest_framework import status
+from rest_framework.authentication import SessionAuthentication
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+
+class Saml2Login(APIView):
+    authentication_classes = (SessionAuthentication,)
+    keys = ['username', 'first_name', 'last_name', 'is_superuser']
+
+    def _prefix(self, keys):
+        return ['{}_saml2_'.format(get_schemaname()) + key for key in keys]
+
+    def _remove_prefix(self, keys):
+        new_keys = dict()
+
+        for k, v in keys.items():
+            new_keys[k.split('{}_saml2_'.format(get_schemaname()))[1]] = v
+
+        return new_keys
+
+    def get(self, request):
+        result = cache.get_many(self._prefix(self.keys))
+
+        return Response(self._remove_prefix(result))
+
+    def delete(self, request):
+        cache.delete_many(self._prefix(self.keys))
+
+        return Response(status=status.HTTP_204_NO_CONTENT)
